add: section the CBCL form into small subforms
